Separate UserStatus-related schema into social content schema.
This makes more semantic sense than storing UserStatus-related type
definitions in the user session schema.
