+++ graphene-django-backend/project/gql_platform/graphene/schema.py
@@ -3,15 +3,18 @@ import graphene
 from graphene_django.types import DjangoObjectType
 
 from config.logger_import import logger
-from project.gql_platform.graphene import user_sessions
+from project.gql_platform.graphene import social_content, user_sessions
 
 
 ''' Query type definition for GraphQL server '''
-class Query(user_sessions.Query, graphene.ObjectType):
+class Query(
+    user_sessions.Query,
+    social_content.Query,
+    graphene.ObjectType):
     pass
 
 ''' Mutation type definition for GraphQL server '''
-class Mutation(user_sessions.Mutation, graphene.ObjectType):
+class Mutation(social_content.Mutation, graphene.ObjectType):
     pass
 
 schema = graphene.Schema(query=Query, mutation=Mutation)
