+++ graphene-django-backend/project/gql_platform/graphene/social_content/__init__.py
@@ -1,0 +1,2 @@
+from .query import Query
+from .mutation import Mutation
