+++ graphene-django-backend/project/gql_platform/graphene/user_sessions/__init__.py
@@ -1,2 +1,1 @@
 from .query import Query
-from .mutation import Mutation
